requirements.txt
fastapi==0.125.0
uvicorn[standard]==0.38.0
jinja2==3.1.6
python-multipart==0.0.21
pydantic==2.12.5
pydantic-settings==2.7.1
docxtpl==0.20.2
python-docx==1.2.0
openpyxl==3.1.5
lxml==5.3.0
supabase==2.15.0
python-dotenv==1.1.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
